Refactor centre determination out into Shifter
This allows MeanShift to not know anything about the internal
representation of values and centres in the Shifter - meaning it is now
possible for clients to implement a custom Shifter without requiring
meanshift to export the data types used internally in the two Shifters
that are implemented here.

TODO(kortschak) Merge the data held by Center and Indices. This will
change Center to interface { Point; Cluster() Indices } and remove the
Clusters() method from Clusterer.